fixed Tests

+++ KeepThingsAPITests/TestModels/TestChat.cs
@@ -2,8 +2,8 @@
 {
     public class TestChat
     {
-        public long sender_id { get; set; }
-        public long receiver_id { get; set; }
+        public int sender_id { get; set; }
+        public int receiver_id { get; set; }
         public string topic { get; set; }
     }
 }
