+++ KeepThingsAPITests/TestModels/TestMessage.cs
@@ -3,7 +3,7 @@ namespace KeepThingsAPITests.Models{
     {
         public int chat_id { get; set; }      
         public string message {get;set;}
-        public long sender_id { get; set; }
+        public int sender_id { get; set; }
         public long sent_timestamp {get;set;}
     }
 }
