+++ KeepThingsAPITests/TestModels/TestUserItem.cs
@@ -3,7 +3,7 @@ namespace KeepThingsAPITests.Models{
     {
         public string item_name {get;set;}
         public string item_desc {get;set;}
-        public long user_id {get;set;}
+        public int user_id {get;set;}
         public string borrower {get;set;}
         public string date_from {get;set;}
         public string date_to {get;set;}
